backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "firebase-admin>=6.7.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
    "sqlmodel>=0.0.21",
    "alembic>=1.13.2",
    "psycopg2-binary>=2.9.9", # Required by Alembic even for SQLite
    "python-multipart>=0.0.20",
]
